Add flag to workaround GHC heisenbug on CI
In very specific tests in GitLab CI we are affected by GHC bug #19421.
We can work around the issue by passing `-with-rtsopts=-xm20000000` when
compiling an affected binary. This is a stopgap measure until the real
bug is fixed.

We have seen the bug in:
- In `clash-testsuite` in `clashLibTest`s
- In `ffi:example` in the `clash` binary itself
- In `prelude:doctests`, probably in the `doctests` binary itself,
  although this is not certain.

This workaround was applied only to those cases that were observed to go
wrong, although as a consequence now the `clash` binary is always built
with the RTS option.