Fix `ffi:example` CI
It did not check the exit codes, continuing after an error and
ultimately reporting success.

`ffi:example` does not profit from having the `build` job's artifact
available, so we remove that dependency. Now `ffi:example` can run
earlier in the CI pipeline.

`ffi:interface-tests` is moved to public runners as it runs in a single
minute.