Add flag to workaround GHC heisenbug on CI (#2444)
In very specific tests in GitLab CI we are affected by GHC bug #19421.
We can work around the issue by passing `-with-rtsopts=-xm20000000` when
compiling an affected binary. This is a stopgap measure until the real
bug is fixed.

We have seen the bug:
- In `clash-testsuite` in `clashLibTest`s
- In `ffi:example` in the `clash` binary itself
- In `prelude:doctests`, probably in the `doctests` binary itself,
  although this is not certain.

and then only in GHC 9.0.2, although the bug should be in other versions
of GHC as well.

This workaround was applied only to GHC 9.0.2 on CI and only to those
cases that were observed to go wrong, although as a consequence now the
`clash` binary is always built with the RTS option.

The `ffi:example` test should no longer be affected by the GHC bug, but
it is a fairly useless test in its current state as it does not fail
when there are issues, and the tested code is actually faulty. It is
disabled to be fixed later.