Evaluate heap dereference in strict contexts
When do we need to force the compile-time evaluation of a prim?:

1. When they are subject of a case-expression
2. When they appear in the argument position of another prim;
   because prims are assumed to be strict

When a prim was in the context of a `Tick`, we already traversed
the context further to determine whether situation 1 or 2 applied.

However, we did not traverse the context further when we were in
the context of a heap update; and that was wrong.

The heap update simply indicates that we dereferend a heap value,
(one that should be updated with the WHNF value once we're done).
But when this variable reference appeared in a situation such as
1 or 2, then heap-bound prim that was pointed to by this reference
must be evaluated to WHNF.