Enable `reduceNonRepPrim` for non-saturated prims
On version 9.6 of GHC, there seems to be less eta-expanding. As
a result, `reduceNonRepPrim` wasn't firing during the
`inlineAndPropagate` stage of the normalization phase as Clash's
eta-expansion only happens later.

This in time made `reduceNonRepPrim` fire at too late a stage,
and certain tests in the regression suite started to fail.

As moving up Clash's eta-expansion isn't an option due to reasons,
the only way forward was for `reduceNonRepPrim` to also fire on
non-saturated prims.