Add custom naming to VIO probes
A design with VIO probes is inspected (through the GUI of the
synthesis tool) when programmed on an FPGA. It therefore makes sense
to give each individual probe a logical name.

The names of in- and output ports should be given as two `Vec`tor of
`Strings`, where the length should match the number of probes.