Simplify GitHub CI matrix, Stack versions
The `.ci` directory contains `stack.yaml` files for each GHC version we
test with Stack on GitHub CI. These files used to be named after the
full GHC version, which is annoying if you want to bump to a later minor
version. The files are now named only after the major GHC version so the
file doesn't change name on such a bump.

The matrix is functionally identical to the old, fully written-out
version.

There is also a bug fix for Windows: we tried to invoke Bash scripts
through PowerShell, the default shell on Windows. This apparently was a
no-op (a failure would have been so much better). These no-ops did not
prevent a succesful run until this commit, it just was not run the way
we thought. We now explicitly specify the shell to use as Bash. Bash is
the default shell on macOS and Linux.