Verilog `tbClockGen`: print nothing when clock stops (#2519)
This ensures that Clash's testsuite doesn't trip up on simulation
finish during normal operation. The testsuite assumes an error
occured when VVP prints anything to stdout. This is needed for
iverilog version 11 and later.

This commit was originally part of #2274 on `master`, but is needed in
1.6 as well.

(cherry picked from commit 07073588403ff39291447a68c7b0bf8891e10f53)

Co-authored-by: Christiaan Baaij <christiaan.baaij@gmail.com>