New: DiffClock
A new type is introduced to signify a differential clock signal that is
passed to the design on two ports in antiphase. This is used by Xilinx
`clockWizardDifferential` (note that Intel solves this case differently,
so if we were to add an Intel primitive it would not need this type).

With this new type, we prevent the creation of more `create_clock`
statements in the global SDC file than is correct: only the positive
phase of a differential pair should have a `create_clock` statement.

Fixes #2518