Make `imap_go` eval rule lazy in index argument (#2543)
Fixes #2542

The problem was that in the final recursive call, when `imap_go`
would return `Nil`, the index argument was out-of-bounds. And
because `imap_go` was strict in all arguments, `imap_go` would
make the `Nil` undefined.

Subsequently applying `reverse` where the `Nil` is undefined,
gives us a complete vector that is undefined, given that
`reverse` is strict in its vector argument.