Use `setName` for `ClockGen` functions
The wizards in `Clash.Xilinx.ClockGen` now use the user-provided name as
the name of the /instance/ rather than the name of the /IP core/. This
change was also done for Intel in PR #1022. When the user is responsible
for creating the IP core, it makes sense to always set the component
name to the user-provided value. But when that is also generated by
Clash, that is no longer needed. Allowing users to set the instance name
makes it possible to match on the instance in SDC files and such.

Instead of always needing a user-specified name for the Intel and Xilinx
PLL functions, the instance name can now be set through
`Clash.Magic.setName`. To accomodate the changed function arguments, the
functions are split into the deprecated old interface and a new
interface which marks the use _unsafe_. This is expanded upon in the
next commits.