Make three reset glitch filter variants (#2544)
The existing code for `resetGlitchFilter` was wrong: the `error` on
domains with unknown initial values rendered as an X value in HDL
instead of throwing an error.

The new `unsafeResetGlitchFilter` allows to sidestep the restriction if
the user is sure this works for them.

The new `resetGlitchFilterWithReset` is the better solution for domains
with unknown initial values, relying on an additional power-on reset
for which assertion glitches don't occur.

Both `resetSynchronizer` and `resetGlitchFilter` no longer carry a
`NOINLINE`/`OPAQUE` annotation. The only reason they had one was for
nicer port/signal names, there was no technical reason to put them in a
separate HDL file. Furthermore, somehow the names for
`resetSynchronizer` were not nice anyway, it would need
`-fno-do-lambda-eta-expansion` to work correctly.