Use DSL for Intel PLL's, account for `ResetPolarity`
`altpll` and `alteraPll` did not account for the input domain's reset
polarity. By rewriting the template functions in the DSL, we can simply
use `DSL.unsafeToActiveHigh` to deal with this.

A new function `Clash.Primitives.DSL.declareN` is added. By expressing
that in terms of `Clash.Netlist.Id.nextN`, it will in the future benefit
from a more efficient implementation when `nextN` is implemented in a
more efficient manner.