Add KnownDomain to unsafeToReset
In preparation for #2589 "Remove KnownDomain".
This is the only constraint addition in clash-prelude in for the KnownDomain removal,
by adding it here for 1.8 we can do the KnownDomain removal later,
with less backwards compatibility issues.

This commit also had to add a KnownDomain to invertReset,
because it used unsafeToReset, but this will be removed again by #2589.