Change Signal.fromList's tail from errorX to error
By making the undefined tail of a signal created with fromList
into an error instead of errorX, this makes it clear when
you're trying use a signal without enough input data.

A problem with the errorX was that in certain circumstances it can be
turned into a signal full of XException. That can turn into a signal
full of undefined BitVectors.
And when that is used as the basis for the expected values
of a outputVerifierBitVector you end up with a testbench that reports
everything is fine for some (possibly big) part of the test.

Also add HasCallStack to Signal.fromList to help with tracing this error.