GitHub CI: Make retrying on runner failure the default
We used to specify retrying for by far most of the jobs, but it was
crucially missing on `nix-build`. So when a runner is shut down during
this job, it will not be retried.

To simplify matters, I just made retrying on runner failure the default.

The only jobs that were not retried before this commit were:
  - nix-build
  - set_pending
  - set_success
  - set_failure

I don't think it hurts to turn it on for those status jobs as well.

Additionally, the `.hackage` inclusion snippet redundantly specified
`{retry: {max: 2}}`, which I removed. Note that GitHub's merge algorithm
for inclusion would fill in the `{retry: {when: ...}}` key from
`.common`, it truly was a no-op in `.hackage`.