DEC: improve interestingPrims detection
For multiplication prims he old code had a handmade distiction between
looking at all arguments, or looking at all but the first argument.
But this doesn't work correctly anymore for some prims.

For example (nowadays?) we have:
  Signed.times# :: Signed m -> Signed n -> Signed (m + n)
  BitVector.times# :: (KnownNat m, KnownNat n) => BitVector m -> BitVector n -> BitVector (m + n)
But the old code used used tailNonPow2 for both.
So for Signed it would only check the last argument.
And for BitVector it would (attempt to) check the `KnownNat n` too.

The new code only has two options:
 * lastNotPow2, check only the last argument, used for div-like prims
 * bothNotPow2, check the last two arguments, used for multiplication prims