Improve (System)Verilog blackbox for Signed.fromInteger#
Make it do the same thing as the VHDL version.
When expanding, it now duplicates the sign bit, preserving the value,
that's new in this commit.
When shrinking it does a bit slice, implementing the wrap-around
behavior that the Haskell simulation does.

(cherry picked from commit d4c5588042dbdc64019bf72edf3f8585f8d45afa)