Fix DDR output primitives
`Clash.Explicit.DDR.ddrOut`: VHDL: Fix incorrect usage of `Enable` signal. With
asynchronous resets, the `Enable` was not connected to the registers.
Furthermore, the `Enable` was combinatorially(!) connected to the mux
that selects the output. The mux should not be affected by the `Enable`
at all (with the current Haskell simulation model).

`Clash.Xilinx.DDR`:
  - `oddr`: Verilog correctly picked one of the possible arguments to
    inspect for `ResetKind`, but the others picked the `KnownNat`
    constraint. This caused Clash to error out during HDL generation.
  - Both in and out prims: Renamed symbols that were incorrectly
    copy-pasted from the Intel prims.

`Clash.Intel.DDR`:
  - `altddioOut`: VHDL: Several numeric identifiers were incorrect,
    referring to wrong arguments or symbols. HDL generation completed
    without error, but the resulting VHDL was completely broken.
  - Both prims: evaluate the `SSymbol` in Haskell to fix HDL generation
    warning about using an argument that is unused in Haskell.

Note that the DDR input primitives were not checked to work correctly
(but the next commit adds a test bench for Xilinx iddr). This PR purely
fixes stuff related to the /output/ primitives, with really minor
adjustments to the input primitives.