No inlining let-bound global vars with clock types (#2846)
The global vars are usually backed by a clock generator that
are not work-free.

In addition, when these global vars are recursively defined,
they can mess up the post-normalization flattening stage which
then violates certain invariants of the netlist generation stage.
This then causes the netlist generation stage to generate bad
Verilog names.

Fixes #2845

(The PR on master was malformed somehow, with the correct contents but
missing the commit message above and an incorrect Git structure. This
backport to 1.8 is reconstructed by hand.)