DDR: document, error on invalid domain
The primitives for Xilinx and Intel only support the rising edge as the
active edge of the domain. To prevent an API change, this is handled
with `clashCompileError`. Clash 1.10 will change the API to enforce it
at the type level.

The added `forall`s do not change the order of the type arguments
(again, avoid an API change).