DDR: Improve API, document
The existing constraint can be redefined in terms of `DomainPeriod`.
Since the virtual DDR domain only exists in simulation, it really
doesn't matter if it doesn't match with the real domain in all its
parameters, so we can drop those and make the constraints more intuitive
to read.

The primitives for Xilinx and Intel only support the rising edge as the
active edge of the domain.

The real domain of the functions was the `slow` domain, and the `fast`
domain was the virtual domain. The new names `dom` and `domDDR`
reinforce the fact that `dom` is the real domain.