Use Word64 Unique on GHC 9.10+
This fixes a left-over from the GHC 9.10 upgrade where we kept
Clash's Unique to be an Int, while GHC's Unique became a Word64

See https://github.com/clash-lang/clash-compiler/pull/2758

Now there is no longer a possibility for overflow when converting
a GHC Unique to a Clash Unique.

As part of this PR we copy the source from the package
https://hackage.haskell.org/package/concurrent-supply
and have it generate Clash Unique values instead of Int.