Update community links in package READMEs (#2896)
These READMEs are rendered as part of the documentation on Hackage, and
are thus pretty prominent. But the links at the bottom aren't very up
to date anymore. I just copied the essence from clash-lang.org.

The `Hackage Dependencies` badge is unfortunately no longer supported:

https://github.com/badges/shields/pull/10618

so we should drop the badge.