Vivado also supports SystemVerilog
Something prevented us from adding this before, but I don't fully
recall.

It works fine now, so let's add it.

One test is affected by #2932. The same test has always had Verilator
excluded, but it turns out Verilator is also affected by #2932, which is
SystemVerilog-specific. Verilator runs fine on the Verilog, so that test
is added.

Note that ModelSim is not affected by #2932 and runs succesfully.