README: Replace bare Unicode
The next commit adds a fourth footnote. Depending on the font, the
superscript numerals for 1-3 render differently than for other numerals
(with 0 possibly rendering in yet another way). By using `<sup>`, we get
consistent rendering for all numbers (and we can add punctuation).

As long as we're making things ugly with the `<sup>` markup, we replace
the emoji-style glyphs with character references. The large glyphs break
many editors, making editing nearly impossible.

This commit does not change the contents, just the way it is written in
source-form.