Use `data` for `Signed`, `Unsigned`, `Index`, and `BiSignalOut`
GHC may insert coercions making Clash see underlying types. For the
types mentioned in the commit title, this means Clash will see either a
`Natural` or an `Integer`. These types are (for better or for worse)
represented as machine words, potentially truncating any operations. As
a concrete example, take:

    top :: Unsinged 68 -> Unsigned 68 -> Vec 1 (Unsigned 64)
    top x y = let res = x + y in res `seq` res :> Nil

Translated to core, we get:

    λ(x :: Unsigned 68) ->
    λ(y :: Unsigned 68) ->
    let
      nt :: Natural
      = +# @68 topEntity2[GlobalId] x[LocalId]
          y[LocalId]
    in case nt[LocalId] of
         _ ->
           Cons @(+ 0 1) @(Unsigned 68) @0
             (_CO_ @(~# Natural Natural (+ 0 1) (+ 0 1)))
             nt[LocalId]
             ((Λb ->
                Nil @0 @b (_CO_ @(~# Natural Natural 0 0)))
                @(Unsigned 68))

I.e., the result of `x + y` gets assinged to a binder of type `Natural`,
making the calculation truncate.