prettyBlackBox: Correctly escape `[` and `]`
The way that way that blackboxes of function arguments to higher order
blackboxes are handled is somewhat weird.

For historical reasons, they are parsed, pretty-printed, later re-parsed
again. However, the pretty-printer failed to correctly re-escape
previously escaped square brackets, resulting in the generation of
invalid HDL, as seen in #2809.

Fixes #2809.
Fixes #2988.