Add `PolyKinds` as default extension to `clash-prelude`
`NoPolyKinds` has confused us multiple times because it would needlessly
constrain instances to `SomeClass (t :: Type)` instead of the more
general `SomeClass (t :: k)`. See the changelog in this commit for more
information.

This also greatly simplifies the reproducer for #1242, as we couldn't
get the behavior we wanted with a local `BitPack` definition --
necessary because instances such as the ones in T1242 are not possible
anymore. The current code in T1242 is built to make sure Clash sees an
unresolved type family which it *has* to resolve in order to produce
HDL. The reproducer was tested by manually reverting the patch for #1242.