Use newer cabal-install on CI
This fixes our documentation rendering issues due to
https://github.com/haskell/cabal/issues/10782

This is a backport of #3035 for 1.8, with some housekeeping added.

Pick up the `CLASH_DOCKER_TAG` and `CACHE_BUST_TOKEN` from PR #2805
(PR #3003 already picked up the env vars).

Change the `CACHE_FALLBACK_KEY` to fall back to the 1.8 branch cache
(which is built nightly) instead of the master branch cache (which will
never exist because the `CLASH_DOCKER_TAG` is always different!).

We can bump the cabal-install versions for all GHC's; #3035 omitted the
one for GHC 8.10 due to a misunderstanding.

Also, the Ubuntu base Docker image is bumped to the latest.