Backport flake.nix changes to 1.8 (#3060)
Cherry picked commits related to the Nix flake from master to create a more suitable environment for downstream projects and mergifybot to work on.

It also updates the flake.lock for newer versions of GHC. This in turn required a patch for clash-cores, similar to the one in the clash-cores repository.