`pack` for `Vec` and `RTree`: more defined bits
This makes `pack` more eager to output defined bits where it can. HDL
already had this behavior, but Clash simulation would output all
undefined bits when the spine was undefined, even though there were
defined elements.

`pack` for `RTree` never needed the `packXWith` in the first place:
`pack` for `Vec` will never return an `XException`, so `pack . t2v`
already never returned an `XException`.

`pack` for `Vec` now no longer needs `packXWith` either: the `pack`ed
elements should not raise an `XException` and the spine is now fully
defined so it also does not raise `XException`.

Omitting `packXWith` is a performance increase (though I don't know how
pronounced, maybe just minor).

The use of `lazyV` and `lazyT` in this instance does not lead to more
laziness. Every constructor of the structure will be walked. On every
element `e` in a vector, `pack e` is evaluated, which results in a
`BV mask val` (`mask` and `val` being a strict `!Natural`). `mask` and
`val` are actually evaluated and used in the result of `pack` for `Vec`
and `RTree`, so the strictness of the fields does not even matter here.

Since every `pack e` call is reduced to `BV mask val`, and `mask` and
`val` are reduced to actual `Natural`s, the `BV mask val` cannot keep a
reference to `e`, meaning the thunk created for `e` by `lazyV` or
`lazyT` is either evaluated or discarded. The vast majority of instances
of `BitPack` will evaluate `e`, but for instance `BitPack ()` does not:
it just reduces to `BV 0 0` regardless of `e`. But in every instance, no
reference to `e` is held by the result of `pack` for `Vec` and `RTree`,
which means no laziness was introduced.

(cherry picked from commit 616d33c41d63e5c82fbd0183b6202f1eff653329)