`pack` should never raise `XException`
It can be nice to be able to rely on `pack` never raising `XException`
and it is good behavior. I have always expected this behavior, but it
was not documented. Let's document it. I have checked all instances in
`clash-prelude` and they all conform to the law. The same goes for
deriving a `BitPack` or using `deriveBitPack`.