Create unit test for `dumpVCD`
Check timing of individual signals when tracing multiple domains

If we have multiple signals, each with their own period, that change
every sample, does the resulting VCD contain all the samples at the
right times, and do timestamps and value changes strictly alternate?

This test would have caught the issue that PR #3076 fixed, and serves as
a good basic sanity check of VCD generation.

A set of ADT's is added that represent all parts of a VCD file that
contain non-trivial data. The ADT's are based on the specified VCD file
syntax in IEEE Std. 1364-2001 (the Verilog standard), which defines the
structure of VCD files. The ADT's (and their naming) correspond to the
names and definitions in section 18.2.1 of the standard; however the
ADT's are simplified to only those parts we use, and folding some parts
into other parts. However, given the basic structure of the `VCDFile`
type, we can trivially extend the types to cover more if we add more
functionality. This is also why we keep the `[DeclarationCommand]` in
`VCDFile` even though what we currently output is always fixed and
direct fields make more sense from that point of view.

`dumpVCD##` is split in such a way that the new `dumpVCD1#` constructs a
`VCDFile` structure and `dumpVCD0#` turns this into a VCD text file.
This allows us to unit test `dumpVCD1#` because all interesting
behavior is there and we don't need to parse a VCD text file to check
for correctness.