deriveAutoReg: Filter NFDataX constraints implied by AutoReg constraints
The previous commit changed the AutoReg instance for a 2-tuple from:
  (AutoReg a, AutoReg b) => AutoReg (a, b)
to:
  (AutoReg a, AutoReg b, NFDataX a, NFDataX b) => AutoReg (a, b)

But since NFDataX is a superclass of AutoReg, these extra constraints are superfluous.
With this extra filtering step we're back to the original.

This isn't strictly necessary, as GHC will handle the superflous constraints without problems.
But it was cluttering up the documentation.